go.tools/ssa/interp: add several standard library packages to test suite.
- Add math.{Log,Ldexp} to externals
- Test now uses FromArgs
- Scan all initial packages for tests, don't assume Created[0] exists.
  (It doesn't if we import a package that has only in-package tests.)

LGTM=gri
R=gri
CC=golang-codereviews
https://golang.org/cl/62010044

+++ go/ssa/interp/reflect.go
@@ -81,7 +81,7 @@ func ext۰reflect۰rtype۰Bits(fr *frame, args []value) value {
 	if !ok {
 		panic(fmt.Sprintf("reflect.Type.Bits(%T): non-basic type", rt))
 	}
-	return fr.i.sizes.Sizeof(basic) * 8
+	return int(fr.i.sizes.Sizeof(basic)) * 8
 }
 
 func ext۰reflect۰rtype۰Elem(fr *frame, args []value) value {
@@ -135,7 +135,7 @@ func ext۰reflect۰rtype۰Out(fr *frame, args []value) value {
 
 func ext۰reflect۰rtype۰Size(fr *frame, args []value) value {
 	// Signature: func (t reflect.rtype) uintptr
-	return uint64(fr.i.sizes.Sizeof(args[0].(rtype).t))
+	return uintptr(fr.i.sizes.Sizeof(args[0].(rtype).t))
 }
 
 func ext۰reflect۰rtype۰String(fr *frame, args []value) value {
